# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: Tax Intelligence Engine API - Pakistan-focused AI Tax Software
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /uploads:
    get:
      operationId: listUploads
      tags: [uploads]
      summary: List uploads for an engagement
      security:
        - bearerAuth: []
      parameters:
        - name: engagementId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Uploads list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadListResponse"
    post:
      operationId: createUpload
      tags: [uploads]
      summary: Upload a document
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/UploadRequest"
      responses:
        "201":
          description: Upload created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Upload"
  /uploads/{id}:
    get:
      operationId: getUpload
      tags: [uploads]
      summary: Get upload status
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Upload details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Upload"
...
  /rules:
    get:
      operationId: listRules
      tags: [rules]
      summary: List tax rules
      security:
        - bearerAuth: []
      parameters:
        - name: taxType
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Rules list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RuleListResponse"
    post:
      operationId: createRule
      tags: [rules]
      summary: Create a new tax rule
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRuleRequest"
      responses:
        "201":
          description: Rule created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaxRule"
  /rules/{engagementId}/run:
    post:
      operationId: runRules
      tags: [rules]
      summary: Run all rules for an engagement
      security:
        - bearerAuth: []
      parameters:
        - name: engagementId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Rules executed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RuleRunResponse"
...
  /risks/{engagementId}:
    get:
      operationId: listRisks
      tags: [risks]
      summary: Get risk register for engagement
      security:
        - bearerAuth: []
      parameters:
        - name: engagementId
          in: path
          required: true
          schema:
            type: string
        - name: severity
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Risk register
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiskListResponse"
  /risks/{id}/status:
    patch:
      operationId: updateRiskStatus
      tags: [risks]
      summary: Update risk status
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RiskStatusUpdate"
      responses:
        "200":
          description: Risk updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RiskFlag"
  /reviews/{engagementId}/notes:
    get:
      operationId: listReviewNotes
      tags: [reviews]
      summary: Get review notes for engagement
      security:
        - bearerAuth: []
      parameters:
        - name: engagementId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Review notes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewNoteListResponse"
    post:
      operationId: createReviewNote
      tags: [reviews]
      summary: Create review note
      security:
        - bearerAuth: []
      parameters:
        - name: engagementId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewNoteRequest"
      responses:
        "201":
          description: Review note created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewNote"
  /reviews/{engagementId}/approve:
    post:
      operationId: approveEngagement
      tags: [reviews]
      summary: Approve engagement (Partner only)
      security:
        - bearerAuth: []
      parameters:
        - name: engagementId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovalRequest"
      responses:
        "200":
          description: Engagement approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Approval"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Upload:
      type: object
      properties:
        id:
          type: string
        engagementId:
          type: string
        fileName:
          type: string
        fileSize:
          type: integer
        mimeType:
          type: string
        category:
          type: string
        parsedStatus:
          type: string
        validationStatus:
          type: string
        sha256Hash:
          type: string
        uploadedBy:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - engagementId
        - fileName
        - category
        - parsedStatus
    UploadRequest:
      type: object
      properties:
        engagementId:
          type: string
        category:
          type: string
        file:
          type: string
          format: binary
      required:
        - engagementId
        - category
        - file
    UploadListResponse:
      type: object
      properties:
        uploads:
          type: array
          items:
            $ref: "#/components/schemas/Upload"
        total:
          type: integer
      required:
        - uploads
        - total
...
    TaxRule:
      type: object
      properties:
        id:
          type: string
        ruleCode:
          type: string
        title:
          type: string
        taxType:
          type: string
        entityType:
          type: string
        conditionJson:
          type: object
        actionJson:
          type: object
        severity:
          type: string
        materialityThreshold:
          type: number
        evidenceRequired:
          type: boolean
        effectiveFrom:
          type: string
          format: date
        effectiveTo:
          type: string
          format: date
        ruleStatus:
          type: string
        versionNo:
          type: integer
      required:
        - id
        - ruleCode
        - title
        - taxType
        - severity
        - ruleStatus
    CreateRuleRequest:
      type: object
      properties:
        ruleCode:
          type: string
        title:
          type: string
        taxType:
          type: string
        entityType:
          type: string
        conditionJson:
          type: object
        actionJson:
          type: object
        severity:
          type: string
        materialityThreshold:
          type: number
        evidenceRequired:
          type: boolean
        effectiveFrom:
          type: string
          format: date
        effectiveTo:
          type: string
          format: date
      required:
        - ruleCode
        - title
        - taxType
        - severity
    RuleListResponse:
      type: object
      properties:
        rules:
          type: array
          items:
            $ref: "#/components/schemas/TaxRule"
        total:
          type: integer
      required:
        - rules
        - total
    RuleRunResponse:
      type: object
      properties:
        rulesEvaluated:
          type: integer
        riskFlagsCreated:
          type: integer
        adjustmentsCreated:
          type: integer
        summary:
          type: string
      required:
        - rulesEvaluated
...
    RiskFlag:
      type: object
      properties:
        id:
          type: string
        engagementId:
          type: string
        riskCode:
          type: string
        severity:
          type: string
        confidence:
          type: number
        estimatedImpact:
          type: number
        title:
          type: string
        description:
          type: string
        recommendation:
          type: string
        status:
          type: string
        owner:
          type: string
        dueDate:
          type: string
          format: date
        disposition:
          type: string
        sourceModule:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - engagementId
        - severity
        - status
    RiskStatusUpdate:
      type: object
      properties:
        status:
          type: string
        disposition:
          type: string
        owner:
          type: string
        dueDate:
          type: string
          format: date
      required:
        - status
    RiskListResponse:
      type: object
      properties:
        risks:
          type: array
          items:
            $ref: "#/components/schemas/RiskFlag"
        total:
          type: integer
        highCount:
          type: integer
        mediumCount:
          type: integer
        lowCount:
          type: integer
      required:
        - risks
        - total
    ReviewNote:
      type: object
      properties:
        id:
          type: string
        engagementId:
          type: string
        content:
          type: string
        authorId:
          type: string
        authorName:
          type: string
        status:
          type: string
        parentId:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - engagementId
        - content
    CreateReviewNoteRequest:
      type: object
      properties:
        content:
          type: string
        parentId:
          type: string
      required:
        - content
    ReviewNoteListResponse:
      type: object
      properties:
        notes:
          type: array
          items:
            $ref: "#/components/schemas/ReviewNote"
        total:
          type: integer
        openCount:
          type: integer
      required:
        - notes
        - total
    ApprovalRequest:
      type: object
      properties:
        comments:
          type: string
      required: []
    Approval:
      type: object
      properties:
        id:
          type: string
        engagementId:
          type: string
        approvedBy:
          type: string
        approverName:
          type: string
        comments:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - engagementId
        - approvedBy
...